+++ changelog.md
@@ -1,3 +1,13 @@
+# 0.1.1
+
+- All class instances will now have an empty Note in `note` attribute instead of None.
+- Add string representation for Note and EnumItem.
+- Enum instance now acts like list of EnumItems.
+- Add EnumType to use in column.type attribute.
+- Column type is now replaced by EnumType instance if enum with such name is defined.
+- Remove unnecessary ColumnType class.
+- Fix: note definition
+
 # 0.1
 
 - Initial release